+++ Examples/SmokeTest/SmokeTest/SmokeTestApp.swift
@@ -10,6 +10,7 @@ struct SmokeTestApp: App {
                 .setAPIEndpoint("http://localhost:4318")
                 .setServiceName("ios-test")
                 .setDebug(true)
+                .setSessionTimeout(10)
                 .build()
             try Honeycomb.configure(options: options)
         } catch {